pyproject.toml
[project]
name = "compliment-of-the-day"
version = "0.1.0"
description = "A Telegram bot that generates funny compliments based on recent news using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "newsapi-python>=0.2.7",
    "pre-commit>=4.3.0",
    "pydantic>=2.11.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[job-queue]>=22.4",
    "sqlalchemy>=2.0.43",
]
